# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FOODIE-FUND",
    "description": "眾資成城-在你心愛的餐廳成為合夥人"
  },
  "servers": [
    {
      "url": "https://foodiefund.onrender.com",
      "description": "Render"
    },
    {
      "url": "http://127.0.0.1:2330",
      "description": "Local"
    }
  ],
...
  "paths": {
    "/v1/api/auth/sign_up": {
      "post": {
        "tags": [
          "使用者登入驗證"
        ],
        "summary": "會員註冊",
        "description": "會員註冊",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "description": "會員資料",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Lobinda"
                  },
                  "email": {
                    "type": "string",
                    "example": "Lobinda123@test.com"
                  },
                  "photo": {
                    "type": "string",
                    "example": ""
                  },
                  "password": {
                    "type": "string",
                    "description": "至少要8碼",
                    "example": "1q2w3e4r"
                  },
                  "phone": {
                    "type": "string",
                    "description": "可Null",
                    "example": "0987654321"
                  },
                  "address": {
                    "type": "string",
                    "description": "地址",
                    "example": "地球某個角落"
                  },
                  "remarks": {
                    "type": "string",
                    "description": "備註",
                    "example": "FOODIE-FUND讚"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "format": "date",
                    "description": "生日",
                    "example": "2006-08-18"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/api/auth/sign_in": {
      "post": {
        "tags": [
          "使用者登入驗證"
        ],
        "summary": "會員登入",
        "description": "會員登入",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "登入成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJ..........mDWPvJZSxu98W4"
                            },
                            "name": {
                              "type": "string",
                              "example": "Lobinda"
                            },
                            "photo": {
                              "type": "string",
                              "example": ""
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "登入成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJ..........mDWPvJZSxu98W4"
                            },
                            "name": {
                              "type": "string",
                              "example": "Lobinda"
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "e-mail尚未驗證"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "e-mail尚未驗證"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "Lobinda123@test.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "1q2w3e4r"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/api/auth/updatePassword": {
      "patch": {
        "tags": [
          "使用者登入驗證"
        ],
        "summary": "更新密碼",
        "description": "更新密碼",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorizatio",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "true"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJ..........mDWPvJZSxu98W4"
                            },
                            "name": {
                              "type": "string",
                              "example": "Lobinda"
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "true"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "example": "eyJhbGciOiJ..........mDWPvJZSxu98W4"
                            },
                            "name": {
                              "type": "string",
                              "example": "Lobinda"
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error Msg"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "至少要8碼",
                    "example": "1q2w3e4r"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "description": "至少要8碼",
                    "example": "1q2w3e4r"
                  }
                },
                "required": [
                  "password",
                  "confirmPassword"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}